push fe

+++ backend/GraficheVeneziane/src/main/java/graficheveneziane/graficheveneziane/form/ProdottiQuantita.java
@@ -8,5 +8,5 @@ import lombok.Setter;
 @Setter
 public class ProdottiQuantita{
     private Prodotto prodotto;
-    private int quantita;
+    private double quantita;
 }
